Make serializing hash-set tags the same as other filters
In this PR we changed how serialization works for filters that were not in hash-set https://github.com/brave/ad-block/commit/74b167c8b49c33f3f61c15dafa2ffde953a25653 and forgot to update it for how it works in hash-set.  Since each was consistent within its own section though it doesn't matter.  I wanted it to be consistent though so I fixed it here.  I made sure already there are no hash-set only filters with tags being used in the lists so this is safe.

Auditors: @SergeyZhukovsky